Added 'description' configuration for use in ingostatusline.vim
This is a shorter, more identifier-like representation than the helptext; the same as SearchSpecial.vim's 'predicateDescription' framed by the /.../ or ?...? indicator for the search direction.
Factored out s:FixedTabWidth().
Moved "related commands" one shiftwidth to the right to make room for the current largest description + helptext. This formatting also nicely prints on 80-column Vim, with the optional related commands column moving to a second line.